added postman collection

+++ laravel/app/Http/Controllers/API/EmployeeController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers\API;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Employee;
+use App\Http\Requests\EmployeeCreateRequest;
+use App\Http\Requests\EmployeeUpdateRequest;
 
 class EmployeeController extends Controller
 {
@@ -15,20 +17,38 @@ class EmployeeController extends Controller
      */
     public function index()
     {
-        //
-        return Employee::all();
+        // Fetch all Employees
+        $allEmployees = Employee::all();
+
+        // Success Reesponse with all employee data
+        return response()->json([
+            "data" => $allEmployees
+        ]);
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  EmployeeCreateRequest  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(EmployeeCreateRequest $request)
     {
-        //
-        $employee = Employee::create($request);
+        // Get validated request body parameters
+        $validated = $request->validated();
+
+        // Create employee
+        $employee = new Employee();
+        $employee->name = $validated['name'];
+        $employee->email = $validated['email'];
+
+        $employee->save();
+
+        // Success Reesponse with created employee data
+        return response()->json([
+            "message" => "Employee is created successfully.",
+            "data" => $employee
+        ]);
     }
 
     /**
@@ -39,29 +59,45 @@ class EmployeeController extends Controller
      */
     public function show($id)
     {
-        //
+        // Fetch employee by id
         $employee = Employee::find($id);
 
-        return $employee;
+        if(!$employee) {
+            return response()->json('Employee not found', 404);
+        }
+
+        return response()->json([
+            "data" => $employee
+        ]);
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  EmployeeUpdateRequest  $request
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(EmployeeUpdateRequest $request, $id)
     {
-        //
+
+        // Get validated request body parameters
+        $validated = $request->validated();
+
+        // Fetch employee by id
         $employee = Employee::find($id);
 
-        $employee->name = $request->name;
-
-        $employee->email = $request->email;
+        // Update employee from request body parameters
+        $employee->name = $validated['name'];
+        $employee->email = $request['email'];
 
         $employee->save();
+
+        // Success Reesponse with updated employee data
+        return response()->json([
+            "message" => "Employee is updated successfully.",
+            "data" => $employee
+        ]);
 
     }
 
@@ -73,9 +109,20 @@ class EmployeeController extends Controller
      */
     public function destroy($id)
     {
-        //
-        $employee = App\Employee::find($id);
+        // Fetch employee by id
+        $employee = Employee::find($id);
 
+        if(!$employee) {
+            return response()->json('Employee not found', 404);
+        }
+
+        // Delete employee
         $employee->delete();
+
+        // Success Reesponse with deleted employee data
+        return response()->json([
+            "message" => "Employee is deleted successfully.",
+            "data" => $employee
+        ]);
     }
 }
